extracted optional_argument
