+++ tests/include/optional_argument_test_fixture.hpp
@@ -6,7 +6,7 @@
 
 #include <cstring>
 
-using ap::argument::optional_argument;
+using ap::argument::optional;
 using ap::detail::argument_name;
 using ap::detail::c_argument_value_type;
 
@@ -17,76 +17,72 @@ struct optional_argument_test_fixture {
     ~optional_argument_test_fixture() = default;
 
     template <c_argument_value_type T>
-    using value_type = typename optional_argument<T>::value_type;
+    using value_type = typename optional<T>::value_type;
 
     template <c_argument_value_type T>
-    void sut_set_used(optional_argument<T>& sut) const {
+    void sut_set_used(optional<T>& sut) const {
         return sut.mark_used();
     }
 
     template <c_argument_value_type T>
-    bool sut_is_used(const optional_argument<T>& sut) const {
+    bool sut_is_used(const optional<T>& sut) const {
         return sut.is_used();
     }
 
     template <c_argument_value_type T>
-    std::size_t sut_get_nused(const optional_argument<T>& sut) const {
+    std::size_t sut_get_nused(const optional<T>& sut) const {
         return sut.nused();
     }
 
     template <c_argument_value_type T>
-    optional_argument<T>& sut_set_value(optional_argument<T>& sut, const std::string& str_value)
-        const {
+    optional<T>& sut_set_value(optional<T>& sut, const std::string& str_value) const {
         sut.mark_used();
         return sut.set_value(str_value);
     }
 
     template <c_argument_value_type T>
-    optional_argument<T>& sut_set_choices(
-        optional_argument<T>& sut, const std::vector<value_type<T>>& choices
-    ) const {
+    optional<T>& sut_set_choices(optional<T>& sut, const std::vector<value_type<T>>& choices)
+        const {
         return sut.choices(choices);
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] bool sut_has_value(const optional_argument<T>& sut) const {
+    [[nodiscard]] bool sut_has_value(const optional<T>& sut) const {
         return sut.has_value();
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] bool sut_has_parsed_values(const optional_argument<T>& sut) const {
+    [[nodiscard]] bool sut_has_parsed_values(const optional<T>& sut) const {
         return sut.has_parsed_values();
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] std::weak_ordering sut_nvalues_in_range(const optional_argument<T>& sut) const {
+    [[nodiscard]] std::weak_ordering sut_nvalues_in_range(const optional<T>& sut) const {
         return sut.nvalues_in_range();
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] const std::any& sut_get_value(const optional_argument<T>& sut) const {
+    [[nodiscard]] const std::any& sut_get_value(const optional<T>& sut) const {
         return sut.value();
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] const std::vector<std::any>& sut_get_values(const optional_argument<T>& sut
-    ) const {
+    [[nodiscard]] const std::vector<std::any>& sut_get_values(const optional<T>& sut) const {
         return sut.values();
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] const argument_name& sut_get_name(const optional_argument<T>& sut) const {
+    [[nodiscard]] const argument_name& sut_get_name(const optional<T>& sut) const {
         return sut.name();
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] bool sut_is_required(const optional_argument<T>& sut) const {
+    [[nodiscard]] bool sut_is_required(const optional<T>& sut) const {
         return sut.is_required();
     }
 
     template <c_argument_value_type T>
-    [[nodiscard]] const std::optional<std::string>& sut_get_help(const optional_argument<T>& sut
-    ) const {
+    [[nodiscard]] const std::optional<std::string>& sut_get_help(const optional<T>& sut) const {
         return sut.help();
     }
 };
